1. Modified DBC to give scaled speed so that float values can be passed. 2. Changed PID logic in RPM, getting actual speed. 3. Fixed reverse issue by calling run motor in 10Hz and adding delay between reverse sequence. rpm speed is also calculated in 10Hz. 3. Improved driver speed calculation and obstacle avoidance logic with float speeds

+++ projects/lpc40xx_motor/l5_application/periodics/motor.h
@@ -5,4 +5,4 @@
 
 void motor__init(void);
 void motor__turn_servo_by_angle(int degrees);
-void motor__run_dc_motor_by_speed(float speed);
+void motor__run_dc_motor_by_speed(double speed);
